+++ partiql-ast-passes/src/partiql_typer.rs
@@ -1,6 +1,6 @@
 use crate::error::{AstTransformError, AstTransformationError};
 use partiql_ast::ast::{
-    AstNode, AstTypeMap, Bag, Expr, List, Lit, NodeId, Query, QuerySet, Struct,
+    AstNode, AstTypeMap, Bag, Expr, List, Lit, NodeId, Query, QuerySet, Struct, TopLevelQuery,
 };
 use partiql_ast::visit::{Traverse, Visit, Visitor};
 use partiql_catalog::Catalog;
@@ -29,7 +29,7 @@ impl<'c> AstPartiqlTyper<'c> {
 
     pub fn type_nodes(
         mut self,
-        query: &AstNode<Query>,
+        query: &AstNode<TopLevelQuery>,
     ) -> Result<AstTypeMap<PartiqlType>, AstTransformationError> {
         query.visit(&mut self);
         if self.errors.is_empty() {
@@ -68,7 +68,7 @@ impl<'c, 'ast> Visitor<'ast> for AstPartiqlTyper<'c> {
 
     fn enter_query_set(&mut self, _query_set: &'ast QuerySet) -> Traverse {
         match _query_set {
-            QuerySet::SetOp(_) => {
+            QuerySet::BagOp(_) => {
                 todo!()
             }
             QuerySet::Select(_) => {}
@@ -220,9 +220,8 @@ impl<'c, 'ast> Visitor<'ast> for AstPartiqlTyper<'c> {
 mod tests {
     use super::*;
     use assert_matches::assert_matches;
-    use partiql_ast::ast;
-    use partiql_catalog::{PartiqlCatalog, TypeEnvEntry};
-    use partiql_types::{PartiqlType, StructConstraint, StructField, TypeKind};
+    use partiql_catalog::PartiqlCatalog;
+    use partiql_types::{PartiqlType, TypeKind};
 
     #[test]
     fn simple_test() {
@@ -262,10 +261,7 @@ mod tests {
             .expect("Expect successful parse");
 
         let typer = AstPartiqlTyper::new(catalog);
-        if let ast::Expr::Query(q) = parsed.ast.as_ref() {
-            typer.type_nodes(&q)
-        } else {
-            panic!("Typing statement other than `Query` are unsupported")
-        }
+        let q = &parsed.ast;
+        typer.type_nodes(q)
     }
 }